+++ support-frontend/assets/pages/weekly-subscription-checkout/components/weeklyCheckoutForm.jsx
@@ -44,12 +44,15 @@ import PersonalDetails from 'components/subscriptionCheckouts/personalDetails';
 import { PaymentMethodSelector } from 'components/subscriptionCheckouts/paymentMethodSelector';
 import CancellationSection from 'components/subscriptionCheckouts/cancellationSection';
 import { newspaperCountries } from 'helpers/internationalisation/country';
+import { GuardianWeekly } from 'helpers/subscriptions';
+import type { IsoCountry } from 'helpers/internationalisation/country';
 
 
 // ----- Types ----- //
 
 type PropTypes = {|
   ...FormFields,
+  country: IsoCountry,
   signOut: typeof signOut,
   formErrors: FormError<FormField>[],
   submissionError: ErrorReason | null,
@@ -63,6 +66,7 @@ type PropTypes = {|
 function mapStateToProps(state: State) {
   return {
     ...getFormFields(state),
+    country: state.common.internationalisation.countryId,
     formErrors: state.page.checkout.formErrors,
     submissionError: state.page.checkout.submissionError,
     productPrices: state.page.checkout.productPrices,
@@ -188,12 +192,13 @@ function WeeklyCheckoutForm(props: PropTypes) {
             </Rows>
           </FormSection>
           <PaymentMethodSelector
-            countrySupportsDirectDebit
+            country={props.country}
+            product={GuardianWeekly}
             paymentMethod={props.paymentMethod}
             setPaymentMethod={props.setPaymentMethod}
             onPaymentAuthorised={props.onPaymentAuthorised}
+            validationError={firstError('paymentMethod', props.formErrors)}
             submissionError={props.submissionError}
-            payPalEnabled={false}
           />
           <FormSection noBorder>
             <Button aria-label={null} type="submit">Continue to payment</Button>
