+++ support-frontend/assets/pages/weekly-subscription-checkout/weeklySubscriptionCheckoutReducer.js
@@ -23,7 +23,7 @@ import { createUserReducer } from 'helpers/user/userReducer';
 import { type PaymentAuthorisation } from 'helpers/paymentIntegrations/readerRevenueApis';
 import type { ProductPrices } from 'helpers/productPrice/productPrices';
 import { type Title } from 'helpers/user/details';
-import { getUser } from 'helpers/user/getUser';
+import { getUser } from 'helpers/user/user';
 import { buildRegularPaymentRequest, onPaymentAuthorised, showPaymentMethod } from './helpers/paymentProviders';
 import { countrySupportsDirectDebit } from 'helpers/paymentProviders';
 import {
